+++ CHANGES.rst
@@ -25,11 +25,11 @@ Changelog
 
 - Fix ownership of comments. [toutpt]
 
-- Provide 'delete own comments' as a configurable option
-  [gyst]
-
 - Make comments editable.
   [pjstevns, gyst]
+
+- Provide 'Delete comments' permission to handle comments deletion
+  [cekk]
 
 2.2.10 (2013-09-24)
 -------------------